requirements.txt
streamlit==1.54.0
streamlit-cropper==0.3.1
Pillow==12.1.0
tensorflow-cpu==2.15.0
numpy==1.26.4
pyarrow==10.0.0
matplotlib==3.7.0

